+++ docs/changelog.md
@@ -31,6 +31,7 @@ As features stabilize some brief notes about them will accumulate here.
 * Improved: font-config queries now run much faster, resulting in snappier startup on unix systems
 * Fixed: [Hide](config/lua/keyassignment/Hide.md) had no effect on macOS when the titlebar was disabled [#679](https://github.com/wez/wezterm/issues/679)
 * Fixed: key and mouse assignments set via [window:set_config_overrides](config/lua/window/set_config_overrides.md) were not respected. [#656](https://github.com/wez/wezterm/issues/656)
+* Fixed: potential panic when word selecting off top of viewport [#713](https://github.com/wez/wezterm/issues/713)
 
 ### 20210405-110924-a5bb5be8